Spring05 에  ExceptionController 사용 예제 추가함

+++ Spring05_Final/src/main/java/com/gura/spring05/cafe/service/CafeServiceImpl.java
@@ -255,6 +255,13 @@ public class CafeServiceImpl implements CafeService{
 	@Override
 	public void deleteComment(HttpServletRequest request) {
 		int num=Integer.parseInt(request.getParameter("num"));
+		
+		CafeCommentDto dto=cafeCommentDao.getData(num);
+		String id=(String)request.getSession().getAttribute("id");
+		if(!dto.getWriter().equals(id)) {
+			throw new NotDeleteException("남의 댓글 지우면 혼난당!");
+		}
+		
 		cafeCommentDao.delete(num);
 	}
 	//댓글 수정 로직 처리
